+++ imdct.h
@@ -1,15 +1,16 @@
 #pragma once
 
+#include "utility.h"
+
 typedef struct {
 	int Bits;
 	int Size;
-	double Scale;
-	double ImdctPrevious[MAX_FRAME_SAMPLES];
-	double* Window;
-	double* SinTable;
-	double* CosTable;
+	DECIMAL Scale;
+	DECIMAL ImdctPrevious[MAX_FRAME_SAMPLES];
+	DECIMAL* Window;
+	DECIMAL* SinTable;
+	DECIMAL* CosTable;
 } Mdct;
 
 void InitMdct();
-void RunImdct(Mdct* mdct, double* input, double* output);
-
+void RunImdct(Mdct* mdct, DECIMAL* input, DECIMAL* output);
